i guiess it's fixed now??
idk this is stupid. It won't fullscreen if the editor is, and it will if the editor isn't full screen. which sure maybe it's a macos quirk or something. but i had completely different behaviour earlier. before i got it to fullscreen while the editor was fullscreened using the exact same shit. tf